ensure device has touch input to use secondary pointer down on double tap
