+++ src/_Shared.Native/Platforms/WinUI/PointerController.cs
@@ -33,7 +33,6 @@ internal partial class PointerController : INativePointerController
 {
     private bool _isPinching;
     private DateTime _pressedTime;
-    private static bool? s_isTouchDevice;
 
     public void InitializeController(object view)
     {
@@ -124,7 +123,7 @@ internal partial class PointerController : INativePointerController
 
     private void OnPinchSarted(object sender, ManipulationStartedRoutedEventArgs e)
     {
-        if (!IsTouchDevice()) return;
+        if (!NativeHelpers.IsTouchDevice()) return;
 
         _isPinching = true;
     }
@@ -139,49 +138,6 @@ internal partial class PointerController : INativePointerController
     private void OnPinchCompleted(object sender, ManipulationCompletedRoutedEventArgs e) =>
         _isPinching = false;
 
-    private static bool IsTouchDevice()
-    {
-        if (s_isTouchDevice.HasValue)
-            return s_isTouchDevice.Value;
-
-        var isWindowsTouchEnabled = false;
-
-#if WINDOWS
-        isWindowsTouchEnabled = WindowsTouchSupportHelper.IsWindowsTouchEnabled();
-#endif
-
-        var result =
-            OperatingSystem.IsAndroid() ||
-            (OperatingSystem.IsIOS() && !OperatingSystem.IsMacCatalyst()) ||
-            //(OperatingSystem.IsBrowser() && IsTouchSupportedInJs()) || is this needed?
-            (OperatingSystem.IsWindows() && isWindowsTouchEnabled);
-
-        s_isTouchDevice = result;
-
-        return result;
-    }
-
-#if WINDOWS
-    public static class WindowsTouchSupportHelper
-    {
-        private const int SM_MAXIMUMTOUCHES = 95;
-
-        [System.Runtime.InteropServices.DllImport("user32.dll")]
-        private static extern int GetSystemMetrics(int nIndex);
-
-        public static bool IsWindowsTouchEnabled()
-        {
-            try
-            {
-                return GetSystemMetrics(SM_MAXIMUMTOUCHES) > 0;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-    }
-#endif
 }
 
 #endif
